backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
redis==5.0.1
marshmallow==3.20.1
qrcode==7.4.2
python-dotenv==1.0.0
gunicorn==21.2.0
razorpay==1.4.2
